requirements.txt
astropy==7.2.0
astropy-iers-data==0.2026.1.26.0.43.56
numpy==2.4.2
pillow==12.1.0
astroquery
pandas